Remove types.ts files
