--- src/components/Base/Link/types.ts
@@ -1,6 +1,0 @@
-import { PropsWithChildren } from 'react'
-import { LinkProps as ChakraLinkProps } from '@chakra-ui/react'
-
-export type LinkProps = ChakraLinkProps & PropsWithChildren<{
-  href: string;
-}>
