--- src/components/Base/Paragraph/types.ts
@@ -1,5 +1,0 @@
-import { TextProps } from '@chakra-ui/react'
-
-export type ParagraphProps = TextProps & {
-  useColorModeVariant?: boolean;
-}
