--- src/components/Base/Popover/types.ts
@@ -1,6 +1,0 @@
-import { PopoverProps as ChakraPopoverProps } from '@chakra-ui/react'
-
-export type PopoverProps = ChakraPopoverProps & {
-  popoverTextElement: JSX.Element;
-  buttonContent: React.ReactNode,
-}
