--- src/components/ContentList/types.ts
@@ -1,5 +1,0 @@
-import { Content } from 'graphql/schema'
-
-export type ContentListProps = {
-  contentList: Array<Content>
-}
