--- src/components/ImageWithCaptions/types.ts
@@ -1,7 +1,0 @@
-import { StackProps } from '@chakra-ui/react'
-
-export type ImageWithCaptionsProps = StackProps & {
-  imageSrc: string;
-  subCaption: string;
-  mainCaption: string;
-}
