--- src/components/PaginationButton/types.ts
@@ -1,5 +1,0 @@
-import { ButtonProps } from '@chakra-ui/react'
-
-export type PaginationButtonProps = ButtonProps & {
-  showMore: boolean;
-}
