--- src/components/Project/types.ts
@@ -1,17 +1,0 @@
-import { BoxProps } from '@chakra-ui/react'
-
-export type ProjectProps = BoxProps & {
-  title: string;
-  liveUrl: string;
-  githubUrl: string;
-  description: string;
-  mainImageUrl: string;
-  stack: {
-    id: string;
-    language: string;
-    framework?: string;
-    libraries?: Array<string>;
-    databases?: Array<string>;
-    categories: Array<string>;
-  };
-}
