--- src/components/ProjectsList/types.ts
@@ -1,6 +1,0 @@
-import { Project, TransformedStack } from 'graphql/schema'
-
-export type ProjectsListProps = {
-  initialProjects: Array<Project>;
-  transformedStack: TransformedStack
-}
