--- src/components/Timeline/Achievements/types.ts
@@ -1,5 +1,0 @@
-import { Achievement } from 'graphql/schema'
-
-export type AchievementsProps = {
-  achievements: Array<Achievement>
-}
