--- src/components/Timeline/types.ts
@@ -1,5 +1,0 @@
-import { Timeline } from 'graphql/schema'
-
-export type TimelineProps = {
-  timelineList: Array<Timeline>
-}
