--- src/components/ToggleThemeIcon/types.ts
@@ -1,6 +1,0 @@
-import { FlexProps } from '@chakra-ui/react'
-
-export type ToggleThemeIconProps = FlexProps & {
-  title: string;
-  isDarkMode?: boolean;
-}
